Added SecurityRestServiceV2
