+++ kura/org.eclipse.kura.rest.security.provider/OSGI-INF/SecurityRestServiceV1.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
-    Copyright (c) 2023 Eurotech and/or its affiliates and others
+    Copyright (c) 2023, 2024 Eurotech and/or its affiliates and others
   
     This program and the accompanying materials are made
     available under the terms of the Eclipse Public License 2.0
@@ -14,11 +14,11 @@
 
 -->
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true"
-               name="org.eclipse.kura.internal.rest.security.provider.SecurityRestService">
-    <implementation class="org.eclipse.kura.internal.rest.security.provider.SecurityRestService"/>
+               name="org.eclipse.kura.internal.rest.security.provider.SecurityRestServiceV1">
+    <implementation class="org.eclipse.kura.internal.rest.security.provider.SecurityRestServiceV1"/>
 
     <property name="kura.service.pid" type="String"
-              value="org.eclipse.kura.internal.rest.security.provider.SecurityRestService"/>
+              value="org.eclipse.kura.internal.rest.security.provider.SecurityRestServiceV1"/>
 
     <reference interface="org.eclipse.kura.cloudconnection.request.RequestHandlerRegistry"
                bind="bindRequestHandlerRegistry"
@@ -40,11 +40,11 @@
                policy="static"/>
 
     <service>
-        <provide interface="org.eclipse.kura.internal.rest.security.provider.SecurityRestService"/>
+        <provide interface="org.eclipse.kura.internal.rest.security.provider.SecurityRestServiceV1"/>
     </service>
 
     <property name="service.pid"
               type="String"
-              value="org.eclipse.kura.internal.rest.security.provider.SecurityRestService"/>
+              value="org.eclipse.kura.internal.rest.security.provider.SecurityRestServiceV1"/>
 
 </scr:component>
