+++ kura/org.eclipse.kura.rest.security.provider/OSGI-INF/SecurityRestServiceV2.xml
@@ -1,0 +1,50 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+    Copyright (c) 2024 Eurotech and/or its affiliates and others
+  
+    This program and the accompanying materials are made
+    available under the terms of the Eclipse Public License 2.0
+    which is available at https://www.eclipse.org/legal/epl-2.0/
+ 
+    SPDX-License-Identifier: EPL-2.0
+    
+    Contributors:
+     Eurotech
+
+-->
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true"
+               name="org.eclipse.kura.internal.rest.security.provider.SecurityRestServiceV2">
+    <implementation class="org.eclipse.kura.internal.rest.security.provider.SecurityRestServiceV2"/>
+
+    <property name="kura.service.pid" type="String"
+              value="org.eclipse.kura.internal.rest.security.provider.SecurityRestServiceV2"/>
+
+    <reference interface="org.eclipse.kura.cloudconnection.request.RequestHandlerRegistry"
+               bind="bindRequestHandlerRegistry"
+               unbind="unbindRequestHandlerRegistry"
+               cardinality="0..n"
+               name="RequestHandlerRegistry"
+               policy="dynamic"/>
+
+    <reference interface="org.eclipse.kura.security.SecurityService"
+               bind="bindSecurityService"
+               cardinality="0..1"
+               name="SecurityService"
+               policy="static"/>
+
+    <reference interface="org.osgi.service.useradmin.UserAdmin"
+               bind="bindUserAdmin"
+               cardinality="1..1"
+               name="UserAdmin"
+               policy="static"/>
+
+    <service>
+        <provide interface="org.eclipse.kura.internal.rest.security.provider.SecurityRestServiceV2"/>
+    </service>
+
+    <property name="service.pid"
+              type="String"
+              value="org.eclipse.kura.internal.rest.security.provider.SecurityRestServiceV2"/>
+
+</scr:component>
