+++ kura/org.eclipse.kura.rest.security.provider/src/main/java/org/eclipse/kura/internal/rest/security/provider/AbstractRestSecurityService.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2023, 2024 Eurotech and/or its affiliates and others
+ * Copyright (c) 2024 Eurotech and/or its affiliates and others
  *
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
@@ -33,25 +33,18 @@ import javax.ws.rs.container.ContainerRequestContext;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import javax.ws.rs.core.Response.Status;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.nio.charset.StandardCharsets;
 import java.util.Optional;
 
-@Path("security/v1")
-public class SecurityRestService {
+public abstract class AbstractRestSecurityService {
 
-    private static final Logger logger = LoggerFactory.getLogger(SecurityRestService.class);
-    private static final String DEBUG_MESSAGE = "Processing request for method '{}'";
+    protected static final Logger logger = LoggerFactory.getLogger(AbstractRestSecurityService.class);
+    protected static final String DEBUG_MESSAGE = "Processing request for method '{}'";
 
-    private static final String MQTT_APP_ID = "SEC-V1";
-    private static final String REST_ROLE_NAME = "security";
-    private static final String KURA_PERMISSION_REST_ROLE = "kura.permission.rest." + REST_ROLE_NAME;
+    protected static final String REST_ROLE_NAME = "security";
+    protected static final String KURA_PERMISSION_REST_ROLE = "kura.permission.rest." + REST_ROLE_NAME;
 
-    private SecurityService security;
-    private final RequestHandler requestHandler = new JaxRsRequestHandlerProxy(this);
+    protected SecurityService security;
+    protected final RequestHandler requestHandler = new JaxRsRequestHandlerProxy(this);
 
     public void bindSecurityService(SecurityService securityService) {
         this.security = securityService;
@@ -63,19 +56,21 @@ public class SecurityRestService {
 
     public void bindRequestHandlerRegistry(RequestHandlerRegistry registry) {
         try {
-            registry.registerRequestHandler(MQTT_APP_ID, this.requestHandler);
+            registry.registerRequestHandler(getMqttAppId(), this.requestHandler);
         } catch (final Exception e) {
-            logger.warn("Failed to register {} request handler", MQTT_APP_ID, e);
+            logger.warn("Failed to register {} request handler", getMqttAppId(), e);
         }
     }
 
     public void unbindRequestHandlerRegistry(RequestHandlerRegistry registry) {
         try {
-            registry.unregister(MQTT_APP_ID);
+            registry.unregister(getMqttAppId());
         } catch (final Exception e) {
-            logger.warn("Failed to unregister {} request handler", MQTT_APP_ID, e);
+            logger.warn("Failed to unregister {} request handler", getMqttAppId(), e);
         }
     }
+
+    public abstract String getMqttAppId();
 
     /**
      * POST method <br /> This method allows the reload of the security policy's fingerprint
@@ -114,48 +109,6 @@ public class SecurityRestService {
     }
 
     /**
-     * POST method <br /> This method replaces the security policy with the default production one. Then a fingerprint
-     * reload is performed.
-     */
-    @POST
-    @RolesAllowed(REST_ROLE_NAME)
-    @Path("/security-policy/apply-default-production")
-    public Response applyDefaultProductionSecurityPolicy() {
-        try {
-            logger.debug(DEBUG_MESSAGE, "applyDefaultProductionSecurityPolicy");
-
-            this.security.applyDefaultProductionSecurityPolicy();
-            this.security.reloadSecurityPolicyFingerprint();
-            this.security.reloadCommandLineFingerprint();
-        } catch (Exception e) {
-            throw DefaultExceptionHandler.toWebApplicationException(e);
-        }
-
-        return Response.ok().build();
-    }
-
-    /**
-     * POST method <br /> This method replaces the security policy with the provided one. Then a fingerprint reload is
-     * performed.
-     */
-    @POST
-    @RolesAllowed(REST_ROLE_NAME)
-    @Path("/security-policy/apply")
-    public Response applySecurityPolicy(InputStream securityPolicyInputStream) {
-        try {
-            logger.debug(DEBUG_MESSAGE, "applySecurityPolicy");
-
-            this.security.applySecurityPolicy(readSecurityPolicyString(securityPolicyInputStream));
-            this.security.reloadSecurityPolicyFingerprint();
-            this.security.reloadCommandLineFingerprint();
-        } catch (Exception e) {
-            throw DefaultExceptionHandler.toWebApplicationException(e);
-        }
-
-        return Response.ok().build();
-    }
-
-    /**
      * GET method
      *
      * @return true if the debug is permitted. False otherwise.
@@ -168,7 +121,7 @@ public class SecurityRestService {
 
             if (context != null && !Optional.ofNullable(context.getSecurityContext())
                     .filter(c -> c.getUserPrincipal() != null).isPresent()) {
-                throw new WebApplicationException(Status.UNAUTHORIZED);
+                throw new WebApplicationException(Response.Status.UNAUTHORIZED);
             }
 
             logger.debug(DEBUG_MESSAGE, "isDebugEnabled");
@@ -177,26 +130,5 @@ public class SecurityRestService {
             throw DefaultExceptionHandler.toWebApplicationException(e);
         }
 
-    }
-
-    private String readSecurityPolicyString(InputStream securityPolicyInputStream) throws IOException {
-        if (securityPolicyInputStream == null) {
-            throw new IllegalArgumentException("Security Policy cannot be null or empty");
-        }
-        int bytesRead;
-        int chunksRead = 0;
-        ByteArrayOutputStream buffer = new ByteArrayOutputStream();
-        byte[] data = new byte[1024];
-        while ((bytesRead = securityPolicyInputStream.read(data, 0, data.length)) != -1) {
-            if (chunksRead++ > 1024) {
-                throw new IllegalArgumentException("Security policy too large");
-            }
-            buffer.write(data, 0, bytesRead);
-        }
-        buffer.flush();
-        if (buffer.size() == 0) {
-            throw new IllegalArgumentException("Security Policy cannot be null or empty");
-        }
-        return new String(buffer.toByteArray(), StandardCharsets.UTF_8);
     }
 }
